Completed merge of acceptance into bidding. Then moved to agent04.
